+++ mlprogram/builtins/__init__.py
@@ -2,6 +2,7 @@ from mlprogram.builtins.apply import Apply  # noqa
 from mlprogram.builtins.datatypes import Environment  # noqa
 from mlprogram.builtins.functions import Flatten  # noqa
 from mlprogram.builtins.functions import Identity  # noqa
-from mlprogram.builtins.functions import Pick  # noqa
 from mlprogram.builtins.functions import Threshold  # noqa
 from mlprogram.builtins.operators import Add, Div, IntDiv, Mul, Neg, Sub  # noqa
+from mlprogram.builtins.pack import Pack  # noqa
+from mlprogram.builtins.pick import Pick  # noqa
