+++ mlprogram/builtins/functions.py
@@ -1,7 +1,6 @@
-from typing import Any, Callable, Generic, List, Optional, TypeVar
+from typing import Any, Callable, Generic, List, TypeVar
 
 from mlprogram import logging
-from mlprogram.builtins.datatypes import Environment
 
 logger = logging.Logger(__name__)
 
@@ -35,11 +34,3 @@ class Threshold(object):
     def __call__(self, value: float) -> bool:
         out = value >= self.threshold
         return self.dtype(out)
-
-
-class Pick(object):
-    def __init__(self, key: str):
-        self.key = key
-
-    def __call__(self, entry: Environment) -> Optional[Any]:
-        return entry[self.key] if self.key in entry.to_dict() else None
